moving components into separate files to make doc less confusing
